+++ app/assets/config/manifest.js
@@ -4,7 +4,9 @@
 //= link application.js
 //= link charts.js
 //= link analytics.js
+//= link pikaday.js
 //= link frappe-gantt.css
+//= link pikaday.css
 //= link_tree ../images
 //= link_tree ../../javascript .js
 //= link_tree ../../../vendor/javascript .js