+++ app/components/student_form_component.rb
@@ -2,11 +2,12 @@ class StudentFormComponent < ViewComponent::Base
   attr_reader :chapter_groups, :permitted_tags
 
   class ChapterGroups
-    attr_reader :groups, :chapter_display
+    attr_reader :groups, :chapter_display, :organization_id
 
     def initialize(chapter, permitted_groups)
       @groups = permitted_groups.filter { |g| g.chapter_id == chapter.id }.sort_by(&:chapter_group_name_with_full_mlid)
       @chapter_display = "#{chapter.organization_name} - #{chapter.chapter_name} (#{chapter.full_mlid})"
+      @organization_id = chapter.organization_id
     end
   end