+++ app/controllers/api/enrollments_controller.rb
@@ -5,8 +5,12 @@ module Api
     has_scope :by_student, as: :student_id
 
     def index
-      @enrollments = apply_scopes(@api_version == 2 ? policy_scope(Enrollment) : Enrollment).all
-      respond_with @enrollments, include: included_params, meta: { timestamp: Time.zone.now }
+      @enrollments = apply_scopes([2, 3].include?(@api_version) ? policy_scope(Enrollment) : Enrollment).all
+      if @api_version == 3
+        respond_with @enrollments, include: included_params, meta: { timestamp: Time.zone.now }, each_serializer: EnrollmentSerializerV2
+      else
+        respond_with @enrollments, include: included_params, meta: { timestamp: Time.zone.now }, each_serializer: EnrollmentSerializer
+      end
     end
 
     def show