+++ db/migrate/20250307233526_add_random_alpha_numeric_generation.rb
@@ -1,0 +1,21 @@
+class AddRandomAlphaNumericGeneration < ActiveRecord::Migration[7.2]
+  def up
+    execute <<~SQL
+      create or replace function random_alphanumeric_string(length integer) returns text
+      language plpgsql
+      as $$
+      declare value text;
+      begin
+        select array_to_string(array(select substr('ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789',floor(random()*36)::int + 1, 1) from generate_series(1,length)),'') into value;
+        return value;
+      end;
+      $$;
+    SQL
+  end
+
+  def down
+    execute <<~SQL
+      drop function random_alphanumeric_string;
+    SQL
+  end
+end