+++ db/migrate/20250505014040_remove_non_null_from_old_mlid.rb
@@ -1,0 +1,5 @@
+class RemoveNonNullFromOldMlid < ActiveRecord::Migration[7.2]
+  def change
+    change_column_null :students, :old_mlid, true
+  end
+end