+++ spec/models/enrollment_spec.rb
@@ -3,8 +3,8 @@
 # Table name: enrollments
 #
 #  id             :uuid             not null, primary key
-#  active_since   :datetime         not null
-#  inactive_since :datetime
+#  active_since   :date             not null
+#  inactive_since :date
 #  created_at     :datetime         not null
 #  updated_at     :datetime         not null
 #  group_id       :bigint           not null
@@ -31,23 +31,23 @@ RSpec.describe Enrollment, type: :model do
   describe 'triggers' do
     before :each do
       @first_group = create :group
-      @second_group = create :group
+      @second_group = create :group, chapter: @first_group.chapter
       @student = create :student, group: @first_group
       @enrollment = create :enrollment, group: @first_group, student: @student
     end
 
     it 'the inactivity on student changing a group' do
-      @student.update(group: @second_group)
+      @student.update!(group: @second_group)
 
-      expect(@enrollment.reload.inactive_since).to be_within(1.second).of(Time.zone.now)
+      expect(@enrollment.reload.inactive_since).to eq(Time.zone.today)
     end
 
     it 'a new enrollment on student changing a group' do
-      @student.update(group: @second_group)
+      @student.update!(group: @second_group)
 
       @new_enrollment = Enrollment.find_by(student: @student, group: @second_group)
 
-      expect(@new_enrollment.active_since).to be_within(1.second).of(Time.zone.now)
+      expect(@new_enrollment.active_since).to eq(Time.zone.today)
       expect(@new_enrollment.inactive_since).to be nil
     end
   end